Keep reuse regressions in existing JSON Schema tests
